fix ett

+++ src/main/java/org/example/chocostyle_datn/repository/DiaChiRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.DiaChi;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiaChiRepository extends JpaRepository<DiaChi, Long> {
+public interface DiaChiRepository extends JpaRepository<DiaChi, Integer> {
 }
