+++ src/main/java/org/example/chocostyle_datn/repository/HoaDonChiTietRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.HoaDonChiTiet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HoaDonChiTietRepository extends JpaRepository<HoaDonChiTiet,Long> {
+public interface HoaDonChiTietRepository extends JpaRepository<HoaDonChiTiet,Integer> {
 }
