+++ src/main/java/org/example/chocostyle_datn/repository/KhachHangRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.KhachHang;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KhachHangRepository extends JpaRepository<KhachHang,Long> {
+public interface KhachHangRepository extends JpaRepository<KhachHang,Integer> {
 }
