+++ src/main/java/org/example/chocostyle_datn/repository/KichCoRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.KichCo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KichCoRepository extends JpaRepository<KichCo,Long> {
+public interface KichCoRepository extends JpaRepository<KichCo,Integer> {
 }
