+++ src/main/java/org/example/chocostyle_datn/repository/KieuDangRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.KieuDang;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KieuDangRepository extends JpaRepository<KieuDang,Long> {
+public interface KieuDangRepository extends JpaRepository<KieuDang,Integer> {
 }
