+++ src/main/java/org/example/chocostyle_datn/repository/LichSuHoaDonRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.LichSuHoaDon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LichSuHoaDonRepository extends JpaRepository<LichSuHoaDon,Long> {
+public interface LichSuHoaDonRepository extends JpaRepository<LichSuHoaDon,Integer> {
 }
