+++ src/main/java/org/example/chocostyle_datn/repository/LoaiAoRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.LoaiAo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoaiAoRepository extends JpaRepository<LoaiAo,Long> {
+public interface LoaiAoRepository extends JpaRepository<LoaiAo,Integer> {
 }
