+++ src/main/java/org/example/chocostyle_datn/repository/MauSacRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.MauSac;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MauSacRepository extends JpaRepository<MauSac,Long> {
+public interface MauSacRepository extends JpaRepository<MauSac,Integer> {
 }
