+++ src/main/java/org/example/chocostyle_datn/repository/NhanVienRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.NhanVien;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NhanVienRepository extends JpaRepository<NhanVien,Long> {
+public interface NhanVienRepository extends JpaRepository<NhanVien,Integer> {
 }
