+++ src/main/java/org/example/chocostyle_datn/repository/PhuongThucThanhToanRepository.java
@@ -3,5 +3,5 @@ package org.example.chocostyle_datn.repository;
 import org.example.chocostyle_datn.entity.PhuongThucThanhToan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PhuongThucThanhToanRepository extends JpaRepository<PhuongThucThanhToan,Long> {
+public interface PhuongThucThanhToanRepository extends JpaRepository<PhuongThucThanhToan,Integer> {
 }
